# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Rathinam Crackers Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: locationId
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
        - name: channel
          in: query
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
        - name: summary
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReportResponse"
  /v1/reports/stock:
    get:
      operationId: getStockReport
      tags: [reports]
      summary: Get stock report
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: lowStockOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Stock report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockReportResponse"
  /v1/reports/outstanding:
    get:
      operationId: getOutstandingReport
      tags: [reports]
      summary: Get outstanding/ageing report
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: customerType
          in: query
          schema:
            type: string
        - name: summary
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Outstanding report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutstandingReportResponse"
  /v1/reports/commission:
    get:
      operationId: getCommissionReport
      tags: [reports]
      summary: Get commission report
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Commission report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionReportResponse"
  /v1/reports/damage:
    get:
      operationId: getDamageReport
      tags: [reports]
      summary: Damage / write-off report
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema: { type: string }
        - name: dateTo
          in: query
          schema: { type: string }
        - name: locationId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Damage report
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      summary:
                        type: object
                        properties:
                          totalUnits: { type: number }
                          eventCount: { type: integer }
                      events:
                        type: array
                        items:
                          type: object
  /v1/reports/loyalty:
    get:
      operationId: getLoyaltyReport
      tags: [reports]
      summary: Loyalty points report
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema: { type: string }
        - name: dateTo
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Loyalty report
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      summary:
                        type: object
                      topCustomers:
                        type: array
                        items: { type: object }
                      recent:
                        type: array
                        items: { type: object }
  /v1/reports/daybook:
    get:
      operationId: getDaybookReport
      tags: [reports]
      summary: Get day book
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Day book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DaybookResponse"

  # ============================================================
  # NOTIFICATIONS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StockLevel:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        productCode:
          type: string
        variantId:
          type: string
        variantSize:
          type: string
        locationId:
          type: string
        locationName:
          type: string
        currentQty:
          type: integer
        reorderLevel:
          type: integer
        isLow:
          type: boolean
...
    SalesReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            summary:
              type: object
              properties:
                totalRevenue:
                  type: number
                totalInvoices:
                  type: integer
                avgOrderValue:
                  type: number
            byChannel:
              type: array
              items:
                type: object
                properties:
                  channel:
                    type: string
                  revenue:
                    type: number
                  count:
                    type: integer
            items:
              type: array
              items:
                type: object
    StockReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/StockLevel"
    OutstandingReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            summary:
              type: object
              properties:
                total:
                  type: number
                bucket0_30:
                  type: number
                bucket31_60:
                  type: number
                bucket61_90:
                  type: number
                bucket90plus:
                  type: number
            customers:
              type: array
              items:
                type: object
    CommissionReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            agents:
              type: array
              items:
                type: object
                properties:
                  agentId:
                    type: string
                  agentName:
                    type: string
                  totalSales:
                    type: number
                  commission:
                    type: number
    DaybookResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            date:
              type: string
            cashIn:
              type: number
            cashOut:
              type: number
            upiIn:
              type: number
            totalSales:
              type: number
            entries:
              type: array
              items:
                type: object

    # NOTIFICATION SCHEMAS
...